refactor(asr): move Parakeet Unified backend out of Streaming/ into its own folder
The unified backend does both streaming and offline batch, so housing it
under Streaming/ was misleading. Moves the whole backend to
ASR/Parakeet/Unified/ (sibling of SlidingWindow/ and Streaming/), and
renames the shared config UnifiedStreamingConfig -> UnifiedConfig since the
offline batch manager uses it too (the L/C/R window fields are documented
as streaming-only). Splits the streaming windower into its own file.
No behavior change; library builds, windower logic re-verified.
