+++ Sources/FluidAudio/ASR/Parakeet/Unified/UnifiedStreamingWindower.swift
@@ -1,76 +1,5 @@
 import Foundation
 
-/// Encoder weight precision for Parakeet Unified 0.6B.
-///
-/// int8 (per-channel linear symmetric) is the default: identical LibriSpeech
-/// test-clean WER to fp16 (offline 1.83% vs 1.82%, streaming 2.14% vs 2.15%),
-/// identical ANE latency on M-series, and half the download/disk (~565 MB vs
-/// ~1.1 GB per encoder).
-public enum UnifiedEncoderPrecision: String, Sendable, CaseIterable {
-    case int8
-    case fp16
-}
-
-/// Configuration for Parakeet Unified 0.6B chunked-attention streaming.
-///
-/// The unified model streams by re-running its stateless encoder over a
-/// `[left | chunk | right]` audio window whose chunked attention mask was
-/// baked in at CoreML conversion time. Context sizes are expressed in 80 ms
-/// encoder frames (1280 samples @ 16 kHz), matching the conversion pipeline
-/// in mobius `models/stt/parakeet-unified-en-0.6b/coreml`.
-public struct UnifiedStreamingConfig: Sendable {
-    /// Left context in encoder frames (history visible to each chunk).
-    public let leftFrames: Int
-    /// Chunk size in encoder frames (new audio decoded per step).
-    public let chunkFrames: Int
-    /// Right context in encoder frames (look-ahead; adds latency).
-    public let rightFrames: Int
-
-    public let sampleRate: Int
-    /// Samples per encoder frame (80 ms @ 16 kHz).
-    public let frameSamples: Int
-    public let melFeatures: Int
-    public let decoderLayers: Int
-    public let decoderHidden: Int
-    public let blankIdx: Int
-    public let maxSymbolsPerFrame: Int
-
-    /// Default export: [70, 13, 13] = 5.6 s left, 1.04 s chunk, 1.04 s right
-    /// (2.08 s theoretical latency — the model card's best-WER streaming mode).
-    public init(
-        leftFrames: Int = 70,
-        chunkFrames: Int = 13,
-        rightFrames: Int = 13,
-        sampleRate: Int = 16000,
-        frameSamples: Int = 1280,
-        melFeatures: Int = 128,
-        decoderLayers: Int = 2,
-        decoderHidden: Int = 640,
-        blankIdx: Int = 1024,
-        maxSymbolsPerFrame: Int = 10
-    ) {
-        self.leftFrames = leftFrames
-        self.chunkFrames = chunkFrames
-        self.rightFrames = rightFrames
-        self.sampleRate = sampleRate
-        self.frameSamples = frameSamples
-        self.melFeatures = melFeatures
-        self.decoderLayers = decoderLayers
-        self.decoderHidden = decoderHidden
-        self.blankIdx = blankIdx
-        self.maxSymbolsPerFrame = maxSymbolsPerFrame
-    }
-
-    /// Total encoder window in samples (left + chunk + right).
-    public var windowSamples: Int { (leftFrames + chunkFrames + rightFrames) * frameSamples }
-    public var chunkSamples: Int { chunkFrames * frameSamples }
-    public var rightSamples: Int { rightFrames * frameSamples }
-    /// Theoretical latency in milliseconds (chunk + right context).
-    public var latencyMs: Int { (chunkFrames + rightFrames) * frameSamples * 1000 / sampleRate }
-    /// Suffix used in streaming encoder file names (e.g. "70_13_13").
-    public var contextSuffix: String { "\(leftFrames)_\(chunkFrames)_\(rightFrames)" }
-}
-
 /// Pure window/frame bookkeeping for unified chunked streaming.
 ///
 /// Mirrors NeMo's `StreamingBatchedAudioBuffer` inference loop: the first step
@@ -78,8 +7,10 @@ public struct UnifiedStreamingConfig: Sendable {
 /// samples. Each step encodes the last `windowSamples` ending at the consumed
 /// position and decodes every not-yet-decoded encoder frame while holding back
 /// the right context (re-encoded with more future audio next step).
+///
+/// Streaming-only — the offline batch path uses `UnifiedBatchLayout` instead.
 struct UnifiedStreamingWindower {
-    let config: UnifiedStreamingConfig
+    let config: UnifiedConfig
 
     /// Global samples fed to the encoder so far.
     private(set) var consumedSamples: Int = 0
@@ -100,7 +31,7 @@ struct UnifiedStreamingWindower {
         let holdbackFrames: Int
     }
 
-    init(config: UnifiedStreamingConfig) {
+    init(config: UnifiedConfig) {
         self.config = config
     }
